requirements.txt
-e .

dvc
tensorflow
matplotlib
numpy
pandas
tqdm
PyYAML
boto3    ## for aws
joblib
scipy

